버그 fix

+++ src/main/java/com/kimcompany/jangbogbackendver2/Member/Repo/MemberRepo.java
@@ -18,10 +18,10 @@ public interface MemberRepo extends JpaRepository<MemberEntity,Long>, MemberSupp
     @Modifying
     @Transactional
     @Query("UPDATE MemberEntity m SET m.lastLoginDate=:now WHERE m.id=:id")
-    Long updateLoginDate(@Param("now")LocalDateTime now ,@Param("id")Long id);
+    Integer updateLoginDate(@Param("now")LocalDateTime now ,@Param("id")Long id);
 
     @Modifying
     @Query("update MemberEntity m set m.failPwd=:num where m.id=:id")
-    Long updatePwdFail(@Param("num")int num,@Param("id")Long id);
+    Integer updatePwdFail(@Param("num")int num,@Param("id")Long id);
 
 }
